# public/doc/openapi-3-v1.1.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This API gets and orders catalog items from different cloud sources.",
    "version": "1.1.0",
    "title": "Catalog API",
    "contact": {
      "email": "support@redhat.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "security": [
    {
      "BasicAuth": [

      ]
    }
  ],
...
  "paths": {
...
    "/portfolios/{portfolio_id}/share_info": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Fetch share information about this portfolio, the response would include a collection of groups and permissions with each group",
        "operationId": "share_info",
        "description": "Fetch share information about a portfolio",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortfolioID"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of group uuid, name and permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareInfo"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "The Portfolio object not found"
          }
        }
      }
    },
...
    "/orders/{order_id}/cancel": {
      "patch": {
        "tags": [
          "Order"
        ],
        "summary": "Cancels a given order",
        "operationId": "cancelOrder",
        "description": "Returns an updated order.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated order object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://cloud.redhat.com/{basePath}",
      "description": "Production Server",
      "variables": {
        "basePath": {
          "default": "/api/catalog/v1.1"
        }
      }
    },
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "Development Server",
      "variables": {
        "port": {
          "default": "3000"
        },
        "basePath": {
          "default": "/api/catalog/v1.1"
        }
      }
    }
  ],
  "components": {
    "parameters": {
...
      "PortfolioID": {
        "name": "portfolio_id",
        "in": "path",
        "description": "The Portfolio ID",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
...
      "OrderID": {
        "name": "order_id",
        "in": "path",
        "description": "The Order ID",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "$ref": "#/components/schemas/ApiErrorCollection"
      },
...
      "Forbidden": {
        "$ref": "#/components/schemas/ApiErrorCollection"
      },
...
      "Unauthorized": {
        "$ref": "#/components/schemas/ApiErrorCollection"
      }
    },
...
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "description": "The userid/password is needed when accessing this API externally",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "user_id": {
            "type": "string",
            "readOnly": true
          },
          "state": {
            "type": "string",
            "enum": [
              "Created",
              "Approval Pending",
              "Ordered",
              "Failed",
              "Completed",
              "Canceled"
            ],
            "title": "State",
            "description": "Current State of the order."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created at",
            "readOnly": true
          },
          "order_request_sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Order request sent at",
            "description": "The time at which the order request was sent to the Topology Service",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed at",
            "readOnly": true
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "example": "jdoe",
            "readOnly": true
          }
        }
      },
...
      "ShareInfo": {
        "type": "object",
        "properties": {
          "group_uuid": {
            "type": "string",
            "title": "The Group UUID",
            "description": "The Group UUID"
          },
          "group_name": {
            "type": "string",
            "title": "The Group Name",
            "description": "The Group Name"
          },
          "permissions": {
            "type": "array",
            "title": "Permissions",
            "description": "One or more permissions currently applied to this group.",
            "items": {
              "type": "string",
              "enum": [
                "read",
                "update",
                "delete",
                "order"
              ]
            }
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "description": "API Error",
        "properties": {
          "status": {
            "type": "string",
            "description": "HTTP status code",
            "example": "400"
          },
          "details": {
            "type": "string",
            "description": "Error details",
            "example": "Bad Request"
          }
        }
      },
      "ApiErrorCollection": {
        "type": "object",
        "description": "API Error collection",
        "properties": {
          "errors": {
            "type": "array",
            "description": "Error list from the API query",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}